refactor: big time file movement
- juggled files around and seperated it into 2 Panes (Side Pane,
  Messages Pane)
- Change terminology to call 1:1 chats Directs, gives us more space in
  terminology
