+++ src/components/MessagesPane/Utilities/Content/HandshakeContent.tsx
