+++ src/components/MessagesPane/Utilities/Utils/BubbleClassGenerator.tsx
