+++ src/components/MessagesPane/Utilities/Utils/bubble-color-generator.ts
